requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.14.0
google-generativeai>=0.3.0
fpdf>=1.7.2
requests>=2.31.0
python-nmap>=0.7.1
